% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_contamination_weights}
\alias{plot_contamination_weights}
\title{Deprecated: plot contamination weights}
\usage{
plot_contamination_weights(x, ...)
}
\arguments{
\item{x}{An \code{sa_contamination_weights} object.}

\item{...}{Passed to the plotting engine; see
\code{\link[=plot.sa_contamination_weights]{plot.sa_contamination_weights()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{plot_contamination_weights()} is deprecated as of fixes 1.0.0: call
\link[=plot.sa_contamination_weights]{plot()} on the result directly.
}
\seealso{
\code{\link[=plot.sa_contamination_weights]{plot.sa_contamination_weights()}}
}
\keyword{internal}
