% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_es}
\alias{plot_es}
\title{Deprecated: plot an event-study result}
\usage{
plot_es(data, ...)
}
\arguments{
\item{data}{An \code{es_result} object.}

\item{...}{Passed to the plotting engine; see \code{\link[=plot.es_result]{plot.es_result()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{plot_es()} is deprecated as of fixes 1.0.0: call
\link[=plot.es_result]{plot()} on the result directly.
}
\seealso{
\code{\link[=plot.es_result]{plot.es_result()}}
}
\keyword{internal}
