% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_es_interactive}
\alias{plot_es_interactive}
\title{Deprecated: interactive event-study plot}
\usage{
plot_es_interactive(data, ...)
}
\arguments{
\item{data}{An \code{es_result} object.}

\item{...}{Passed to the plotting engine; see \code{\link[=plot.es_result]{plot.es_result()}}.}
}
\value{
A \code{plotly} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{plot_es_interactive()} is deprecated as of fixes 1.0.0: use
\code{plot(x, interactive = TRUE)} instead; see \code{\link[=plot.es_result]{plot.es_result()}}.
}
\seealso{
\code{\link[=plot.es_result]{plot.es_result()}}
}
\keyword{internal}
