% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{run_did}
\alias{run_did}
\title{Deprecated: run a basic TWFE DiD model}
\usage{
run_did(
  data,
  outcome,
  treatment,
  timing = NULL,
  fe = NULL,
  unit = NULL,
  time = NULL,
  covariates = NULL,
  cluster = NULL,
  weights = NULL,
  conf.level = 0.95,
  vcov = "HC1",
  vcov_args = list()
)
}
\arguments{
\item{data, outcome, treatment, timing, fe, unit, time}{Passed to \code{\link[=did]{did()}}.}

\item{covariates, cluster, weights, vcov, vcov_args}{Passed to \code{\link[=did]{did()}}.}

\item{conf.level}{Renamed to \code{conf_level} in \code{\link[=did]{did()}}.}
}
\value{
A \code{did_result} object; see \code{\link[=did]{did()}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{run_did()} was renamed to \code{\link[=did]{did()}} in fixes 1.0.0.
}
\seealso{
\code{\link[=did]{did()}}
}
\keyword{internal}
