% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_study.R
\name{plot_es}
\alias{plot_es}
\title{Plot Event Study Results}
\usage{
plot_es(
  data,
  type = "ribbon",
  vline_val = 0,
  vline_color = "#000",
  hline_val = 0,
  hline_color = "#000",
  linewidth = 1,
  pointsize = 2,
  alpha = 0.2,
  barwidth = 0.2,
  color = "#B25D91FF",
  fill = "#B25D91FF"
)
}
\arguments{
\item{data}{A dataframe containing the results from the `run_es` function.}

\item{type}{The type of confidence interval visualization: "ribbon" (default) or "errorbar".}

\item{vline_val}{The x-intercept for the vertical reference line (default: 0).}

\item{vline_color}{Color for the vertical reference line (default: "#000").}

\item{hline_val}{The y-intercept for the horizontal reference line (default: 0).}

\item{hline_color}{Color for the horizontal reference line (default: "#000").}

\item{linewidth}{The width of the lines for the plot (default: 1).}

\item{pointsize}{The size of the points for the estimates (default: 2).}

\item{alpha}{The transparency level for ribbons (default: 0.2).}

\item{barwidth}{The width of the error bars (default: 0.2).}

\item{color}{The color for the lines and points (default: "#B25D91FF").}

\item{fill}{The fill color for ribbons (default: "#B25D91FF").}
}
\value{
A ggplot object displaying the event study results.
}
\description{
This function creates a plot for event study results using `ggplot2`.
Users can choose between ribbon-style confidence intervals or error bars.
}
